+++ cmd/services.go
@@ -184,6 +184,7 @@ type APIClient interface {
 	GetApplicationDemoConfig(requestContext context.Context, applicationID string) (json.RawMessage, error)
 	UpdateApplicationDemoConfig(requestContext context.Context, applicationID string, configuration json.RawMessage) (json.RawMessage, error)
 	FixApplicationDemo(requestContext context.Context, applicationID string, workspaceID string) (json.RawMessage, error)
+	FixApplicationBuild(requestContext context.Context, applicationID string, branch string) (json.RawMessage, error)
 	GetApplicationImageRegistry(requestContext context.Context, applicationID string) (json.RawMessage, error)
 	UpdateApplicationImageRegistry(requestContext context.Context, applicationID string, registryRequest client.UpdateApplicationImageRegistryRequest) (json.RawMessage, error)
 	GetApplicationAIConfig(requestContext context.Context, applicationID string) (json.RawMessage, error)